Add fixtures & tooling to simplify compile tests
Provide `pip_produces_absolute_paths` as a fixture, removing the need
for some individual tests to check the pip version directly.

The `TestFilesCollection` dataclass is a tool for writing parameters
which includes the ability to populate a temp dir based on the
structured data within. This simplifies test parameters and behavior.

Additionally, mildly cleanup tests to use monkeypatch.context()